Fix: Last Node of each node_chunk getting skipped
The last node of every node_chunk was not sending any messages to its
neighbours (in the exchange_messages kernel function). This was because
the global_clause_node_output, source_node_chunk and source_node_pos
where signed integers, so the if condition was evaluated to false.
Making them unsigned fixes the problem.